added extra fields in models + created user model + Pydantic schemas refactoring

+++ src/migrations/versions/2025_06_12_2041-bb5818568d87_added_rooms.py
@@ -25,7 +25,7 @@ def upgrade() -> None:
         sa.Column("hotel_id", sa.Integer(), nullable=False),
         sa.Column("title", sa.String(), nullable=False),
         sa.Column("description", sa.String(), nullable=True),
-        sa.Column("price", sa.Integer(), nullable=False),
+        sa.Column("price", sa.Float(), nullable=False),
         sa.Column("quantity", sa.Integer(), nullable=False),
         sa.ForeignKeyConstraint(
             ["hotel_id"],
